+++ protected/humhub/modules/user/widgets/PeopleIcons.php
@@ -2,7 +2,7 @@
 
 /**
  * @link https://www.humhub.org/
- * @copyright Copyright (c) 2021 HumHub GmbH & Co. KG
+ * @copyright Copyright (c) HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
  */
 
@@ -14,7 +14,7 @@ use humhub\modules\user\models\User;
 /**
  * PeopleIcons shows footer icons for people cards
  *
- * @since 1.9
+ * @since 1.17.2
  * @author Luke
  */
 class PeopleIcons extends Widget
@@ -25,13 +25,31 @@ class PeopleIcons extends Widget
     public $user;
 
     /**
+     * @var string $separator Separator between icons
+     */
+    public string $separator = ' ';
+
+    /**
+     * @var array $icons An icon can be HTML code or object convertable to string
+     */
+    protected array $icons = [];
+
+    /**
      * @inheritdoc
      */
     public function run()
     {
-        return $this->render('peopleIcons', [
-            'user' => $this->user,
-        ]);
+        return implode($this->separator, $this->icons);
     }
 
+    /**
+     * Add an icon to this widget
+     *
+     * @param mixed $icon HTML code or object convertable to string
+     * @return void
+     */
+    public function addIcon($icon): void
+    {
+        $this->icons[] = $icon;
+    }
 }